dfs replaces bfs
