+++ README.md
@@ -7,7 +7,7 @@ It is also possible to do regression, including multi-output regression.
 
 ## Installation
 ```bash
-pip install git+https://github.com/KohlerHECTOR/dpdt-py.git@v0.1.7
+pip install git+https://github.com/KohlerHECTOR/dpdt-py.git@v0.2.0
 ```
