+++ dpdt/__init__.py
@@ -1,5 +1,1 @@
 from .dpdt_sklearn import DPDTreeClassifier, DPDTreeRegressor
-from .dpdt_sklearn_depth_first import (
-    DPDTreeClassifierDepthFirst,
-    DPDTreeRegressorDepthFirst,
-)
