+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 import os
 
-__version__ = "0.1.7"
+__version__ = "0.2.0"
 
 packages = find_packages(
     exclude=[
